@@ -5,7 +5,7 @@
  */
 package com.portfolio.lucas.Dto;
 
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotBlank;
 
 public class EducacionData {
     @NotBlank
