@@ -4,7 +4,7 @@
  */
 package com.portfolio.lucas.Dto;
 
-import javax.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotBlank;
 
 /**
  *
