@@ -5,8 +5,7 @@
 package com.portfolio.lucas.Security.Service;
 
 import com.portfolio.lucas.Security.Entity.Rol;
-import com.portfolio.lucas.Security.Enums.RolNombre;
-import javax.transaction.Transactional;
+import jakarta.transaction.Transactional;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,7 +19,7 @@ import com.portfolio.lucas.Security.Repository.IRolRepo;
 @Transactional
 public class RolService {
     @Autowired IRolRepo irolRepository;
-    public Optional<Rol> getByRolNombre(RolNombre rolNombre){
+    public Optional<Rol> getByRolNombre(String rolNombre){
         return irolRepository.findByRolNombre(rolNombre);
     }
     public void save(Rol rol){
