@@ -5,7 +5,7 @@
 package com.portfolio.lucas.Security.Service;
 
 import com.portfolio.lucas.Security.Entity.Usuario;
-import javax.transaction.Transactional;
+import jakarta.transaction.Transactional;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
