@@ -7,7 +7,7 @@ package com.portfolio.lucas.Service;
 import com.portfolio.lucas.Entity.Experiencia;
 import java.util.List;
 import java.util.Optional;
-import javax.transaction.Transactional;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.portfolio.lucas.Repository.IExperienciaRepo;
