@@ -5,7 +5,6 @@
 package com.portfolio.lucas.Security.Repository;
 
 import com.portfolio.lucas.Security.Entity.Rol;
-import com.portfolio.lucas.Security.Enums.RolNombre;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -16,5 +15,5 @@ import org.springframework.stereotype.Repository;
  */
 @Repository
 public interface IRolRepo extends JpaRepository<Rol,Integer>{
-    Optional<Rol> findByRolNombre(RolNombre rolNombre);
+    Optional<Rol> findByRolNombre(String rolNombre);
 }
